chore(view): update imports after utility file restructuring
