+++ src/main/kotlin/lotto/view/input/Bonus.kt
@@ -3,7 +3,6 @@ package lotto.view.input
 import lotto.error.ErrorType
 import lotto.policy.isInRange
 
-
 fun parseBonusNumber(input: String): Int? {
 	return input.toIntOrNull()
 }
